requirements.txt
amqp==2.6.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
anyio==3.4.0; python_version >= "3.6" and python_full_version >= "3.6.2"
asgiref==3.4.1; python_full_version >= "3.6.1" and python_version >= "3.6"
async-generator==1.10; python_version >= "3.7" and python_version < "4.0"
async-timeout==4.0.2; python_version >= "3.6"
attrs==21.2.0; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.5.0"
billiard==3.6.4.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
# celery==4.4.7; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
certifi==2021.10.8
cffi==1.15.0; os_name == "nt" and implementation_name != "pypy" and python_version >= "3.7" and python_version < "4.0" and (python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7")
charset-normalizer==2.0.9; python_full_version >= "3.6.1" and python_version >= "3" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0")
click==8.0.3; python_full_version >= "3.6.1" and python_version >= "3.6"
colorama==0.4.4
cryptography==36.0.0; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7"
deprecated==1.2.13; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
dnspython==2.1.0; python_full_version >= "3.6.1" and python_version >= "3.6"
email-validator==1.1.3; python_full_version >= "3.6.1"
fastapi==0.70.1; python_full_version >= "3.6.1"
greenlet==1.1.2; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and python_full_version >= "3.5.0"
h11==0.12.0; python_full_version >= "3.6.1" and python_version >= "3.7" and python_version < "4.0"
httptools==0.2.0
idna==3.3; python_full_version >= "3.6.2" and python_version >= "3.7" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and python_version < "4.0"
itsdangerous==2.0.1; python_full_version >= "3.6.1" and python_version >= "3.6"
jinja2==3.0.3; python_full_version >= "3.6.1" and python_version >= "3.6"
joblib==1.1.0; python_version >= "3.7"
kombu==4.6.11; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
markupsafe==2.0.1; python_full_version >= "3.6.1" and python_version >= "3.6"
mypy-extensions==0.4.3; python_version >= "3.5" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0")
mypy==0.910; python_version >= "3.5"
numpy==1.22.3; python_version >= "3.8"
oauthlib==3.2.2; python_version >= "3.6"
opencv-python==4.5.5.64; python_version >= "3.6"
orjson==3.6.5; python_full_version >= "3.6.1" and python_version >= "3.7"
outcome==1.1.0; python_version >= "3.7" and python_version < "4.0"
packaging==21.3; python_version >= "3.6"
psycopg2-binary==2.9.2; python_version >= "3.6"
pycparser==2.21; python_version >= "3.7" and python_full_version < "3.0.0" and os_name == "nt" and implementation_name != "pypy" and python_version < "4.0" and (python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7") or os_name == "nt" and implementation_name != "pypy" and python_version >= "3.7" and python_version < "4.0" and (python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7") and python_full_version >= "3.4.0"
pydantic==1.8.2; python_full_version >= "3.6.1"
pyopenssl==21.0.0; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7"
pyparsing==3.0.6; python_version >= "3.6"
python-dateutil==2.8.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
python-dotenv==0.19.2; python_version >= "3.5"
python-multipart==0.0.5; python_full_version >= "3.6.1"
python3-discogs-client==2.5
pytz==2022.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pyyaml==5.4.1
redis==4.3.1; python_version >= "3.6"
requests==2.26.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
scikit-learn==1.0.2; python_version >= "3.7"
scipy==1.6.1; python_version >= "3.7"
selenium==4.1.0; python_version >= "3.7" and python_version < "4.0"
six==1.16.0; python_full_version >= "3.6.1" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0") and (python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7")
sklearn==0.0
sniffio==1.2.0; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.2"
sortedcontainers==2.4.0; python_version >= "3.7" and python_version < "4.0"
spotipy==2.19.0
sqlalchemy-stubs==0.4
sqlalchemy2-stubs==0.0.2a19; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
sqlalchemy==1.4.53; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
starlette==0.16.0; python_version >= "3.6" and python_full_version >= "3.6.1"
threadpoolctl==3.1.0; python_version >= "3.7"
toml==0.10.2; python_version >= "3.5" and python_full_version < "3.0.0" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") or python_version >= "3.5" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and python_full_version >= "3.3.0"
trio-websocket==0.9.2; python_version >= "3.7" and python_version < "4.0"
trio==0.19.0; python_version >= "3.7" and python_version < "4.0"
types-requests==2.26.1
typing-extensions==4.0.1; python_version >= "3.6" and python_full_version >= "3.6.1" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
ujson==4.3.0; python_full_version >= "3.6.1" and python_version >= "3.6"
urllib3==1.26.7; python_full_version >= "3.6.1" and python_version < "4" and python_version >= "3.7" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0")
uvicorn==0.15.0; python_full_version >= "3.6.1"
uvloop==0.16.0
vine==1.3.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
watchgod==0.7
websockets==10.1
wrapt==1.14.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
wsproto==1.0.0; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.1"
